requirements.txt
picamera==1.13
pyserial==3.4
requests==2.24.0
urllib3==1.25.9
opencv-python-headless==4.4.0.44
scipy
